+++ cgogn/rendering/mesh_render.h
@@ -171,7 +171,7 @@ protected:
 				auto& vertices = vvertices[worker_index];
 				vertices.clear();
 				incident_vertices(m, f, vertices);
-				for (uint32 i = 1; i < vertices.size() - 1; ++i)
+				for (uint32 i = 1; i < uint32(vertices.size()) - 1; ++i)
 				{
 					auto& tif = table_indices[worker_index];
 					tif.push_back(index_of(m, vertices[0]));
@@ -333,7 +333,7 @@ public:
 		auto func_update_ebo = [&](DrawingType pr, const TablesIndices& table) -> void {
 			uint32 total_size = 0;
 			for (const auto& t : table)
-				total_size += t.size();
+				total_size += uint32(t.size());
 
 			indices_buffers_uptodate_[pr] = true;
 			if (total_size > 0)
@@ -345,8 +345,8 @@ public:
 				uint32 beg = 0;
 				for (const auto& t : table)
 				{
-					indices_buffers_[pr]->copy_data(beg, t.size(), t.data());
-					beg += t.size();
+					indices_buffers_[pr]->copy_data(beg, uint32(t.size()), t.data());
+					beg += uint32(t.size());
 				}
 				indices_buffers_[pr]->set_name("EBO_" + primitives_names[pr]);
 			}
@@ -355,7 +355,7 @@ public:
 		auto func_update_ebo2 = [&](DrawingType pr1, const TablesIndices& table1) -> void {
 			uint32 total_size1 = 0;
 			for (const auto& t : table1)
-				total_size1 += t.size();
+				total_size1 += uint32(t.size());
 
 			indices_buffers_uptodate_[pr1] = true;
 			if (total_size1 > 0)
@@ -383,7 +383,7 @@ public:
 									uint32 interv) -> void {
 			uint32 total_size1 = 0;
 			for (const auto& t : table1)
-				total_size1 += t.size();
+				total_size1 += uint32(t.size());
 
 			indices_buffers_uptodate_[pr1] = true;
 			if (total_size1 > 0)
@@ -401,7 +401,7 @@ public:
 				for (int j = 0; j < nb; ++j)
 				{
 					const auto& t1 = table1[j];
-					uint32 sz = t1.size();
+					uint32 sz = uint32(t1.size());
 					for (uint32 k = 0; k < sz; ++k)
 						*ptr1++ = (k % interv == interv - 1) ? t1[k] + beg : t1[k];
